Refactored subscriber code into different files (for now at least)
